# swagger.yaml
openapi: 3.1.0
info:
  title: MMA Gym API
  description: Complete API for managing MMA gym operations including classes, bookings, shop, and user management
  version: 1.0.0
  contact:
    name: Apex MMA
    email: dev@apexmma.com
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: http://localhost:3001
    description: Development root
  - url: https://api.apexmma.com
    description: Production server
...
paths:
...
  /content/{key}:
    get:
      tags: [Content]
      summary: Get content block by key
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
          example: "home.hero"
        - in: query
          name: locale
          schema:
            type: string
            default: en
      responses:
        '200':
          description: Content block retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      tags: [Content]
      summary: Update content block
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [locale, json]
              properties:
                locale:
                  type: string
                  default: en
                json:
                  type: object
      responses:
        '200':
          description: Content block updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags: [Content]
      summary: Delete content block
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
        - in: query
          name: locale
          schema:
            type: string
            default: en
      responses:
        '200':
          description: Content block deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Content block not found
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        data:
          oneOf:
            - type: object
            - type: array
        meta:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []